include/digest_fe_internal.h: put state union first in context structure
This is by far the largest object and it will ensure that it is aligned
for access by the underlying digest algorithm.

SHA2-512 is a 64-bit algorithm, and if we are building on a 32-bit
machine, the 3 function pointers that were located before it will result
in the state union not being 64-bit aligned.

[ci skip]